+++ src/tno/quantum/utils/__init__.py
@@ -13,6 +13,7 @@ from tno.quantum.utils._base_arguments import BaseArguments
 from tno.quantum.utils._base_config import BaseConfig
 from tno.quantum.utils._bit_vector import BitVector, BitVectorLike
 from tno.quantum.utils._utils import (
+    check_equal,
     convert_to_snake_case,
     get_installed_subclasses,
 )
@@ -22,18 +23,20 @@ __all__ = [
     "BaseConfig",
     "BitVector",
     "BitVectorLike",
+    "check_equal",
     "convert_to_snake_case",
     "get_installed_subclasses",
 ]
 
 if importlib.util.find_spec("pennylane") is not None:
-    from tno.quantum.utils._base_config import BackendConfig
+    from tno.quantum.utils._backend_config import BackendConfig
+    from tno.quantum.utils._noise_config import NoiseConfig
 
-    __all__ += ["BackendConfig"]
+    __all__ += ["BackendConfig", "NoiseConfig"]
 
 if importlib.util.find_spec("torch") is not None:
-    from tno.quantum.utils._base_config import OptimizerConfig
+    from tno.quantum.utils._optimizer_config import OptimizerConfig
 
     __all__ += ["OptimizerConfig"]
 
-__version__ = "5.1.0"
+__version__ = "5.4.1"